Define h5py storage

+++ flow_storage/flow_io_utils_impl/flowioutils_fs.py
@@ -3,26 +3,36 @@ import os, os.path
 import cv2
 import numpy as np
 import json
+
 from typing import Dict, List, Tuple
 
-class FlowIOUtilsFs():
-  def __init__(self) -> None:
-      pass
+from ..flowstorageconfig import FlowStorageConfig
 
-  def clean_ext_storage(self, path: str) -> None:
-    for root, dirs, files in os.walk(path):
+
+class FlowIOUtilsFs():
+  def __init__(self, config: FlowStorageConfig) -> None:
+      self._config = config
+
+  def clean_ext_storage(self) -> None:
+    if self._config.storage_location == '.':
+      print('storage location is not defined!!!')
+      return
+      
+    for root, dirs, files in os.walk(self._config.storage_location):
       for file in files:
         os.remove(os.path.join(root, file))
     return
 
 # Readers
-  def np_array_reader(self, ffn: str) -> np.ndarray:
+  def np_array_reader(self, fn: str) -> np.ndarray:
+    ffn = f'{self._config.storage_location}/{fn}'
     ffn = f'{ffn}.npy'
     if isfile(ffn):
       return np.load(ffn)
     return None
 
-  def json_reader(self, ffn: str) -> Dict:
+  def json_reader(self, fn: str) -> Dict:
+    ffn = f'{self._config.storage_location}/{fn}'
     ffn = f'{ffn}.json'
     if not isfile(ffn):
       return None
@@ -30,7 +40,8 @@ class FlowIOUtilsFs():
       data = json.load(f)
       return data
 
-  def list_np_arrays_reader(self, ffn: str) -> List[np.ndarray]:
+  def list_np_arrays_reader(self, fn: str) -> List[np.ndarray]:
+    ffn = f'{self._config.storage_location}/{fn}'
     ffn = f'{ffn}.json'
     if not isfile(ffn):
       return None
@@ -39,14 +50,15 @@ class FlowIOUtilsFs():
       data = [np.array(d) for d in ld]
       return data
 
-  def list_tuples_reader(self, ffn: str) -> List[Tuple]:
+  def list_tuples_reader(self, fn: str) -> List[Tuple]:
+    ffn = f'{self._config.storage_location}/{fn}'
     ffn = f'{ffn}.json'
     with open(ffn, 'rt') as f:
       ld = json.load(f)
       data = [np.array(d) for d in ld]
       return data
 
-  def list_keypoints_reader(self, ffn: str) -> List[cv2.KeyPoint]:
+  def list_keypoints_reader(self, fn: str) -> List[cv2.KeyPoint]:
 
     def _list_dict_to_list_key_points(data: List[Dict]) -> List[cv2.KeyPoint]:
       list_kps = []
@@ -64,6 +76,7 @@ class FlowIOUtilsFs():
         list_kps.append(kp)
       return list_kps
 
+    ffn = f'{self._config.storage_location}/{fn}'
     ffn = f'{ffn}.json'
     with open(ffn, 'rt') as f:
       data = json.load(f)
@@ -71,31 +84,35 @@ class FlowIOUtilsFs():
       return list_kps
 
 # Writers
-  def np_array_writer(self, ffn: str, arr: np.ndarray) -> None:
+  def np_array_writer(self, fn: str, arr: np.ndarray) -> None:
+    ffn = f'{self._config.storage_location}/{fn}'
     ffn = f'{ffn}.npy'
     np.save(ffn, arr)
     return
 
-  def list_np_arrays_writer(self, ffn: str, data: List[np.ndarray]) -> None:
+  def list_np_arrays_writer(self, fn: str, data: List[np.ndarray]) -> None:
+    ffn = f'{self._config.storage_location}/{fn}'
     ffn = f'{ffn}.json'
     sd = [d.tolist() for d in data]
     with open(ffn, 'w') as fp:
       json.dump(sd, fp, indent=2)
     return
 
-  def json_writer(self, ffn: str, data: Dict) -> None:
+  def json_writer(self, fn: str, data: Dict) -> None:
+    ffn = f'{self._config.storage_location}/{fn}'
     ffn = f'{ffn}.json'
     with open(ffn, 'w') as fp:
       json.dump(data, fp, indent=2)
     return
 
-  def list_tuples_writer(self, ffn: str, data: List[Tuple]) -> None:
+  def list_tuples_writer(self, fn: str, data: List[Tuple]) -> None:
+    ffn = f'{self._config.storage_location}/{fn}'
     ffn = f'{ffn}.json'
     with open(ffn, 'w') as fp:
       json.dump(data, fp, indent=2)
     return
 
-  def list_keypoints_writer(self, ffn: str, data: List[cv2.KeyPoint]) -> None:
+  def list_keypoints_writer(self, fn: str, data: List[cv2.KeyPoint]) -> None:
 
     def _list_key_points_to_List_dict(data: List[cv2.KeyPoint]) -> List[Dict]:
       list_dict = []
@@ -111,6 +128,7 @@ class FlowIOUtilsFs():
         list_dict.append(kp_dict)
       return list_dict
 
+    ffn = f'{self._config.storage_location}/{fn}'
     ffn = f'{ffn}.json'
     kps = _list_key_points_to_List_dict(data)
     with open(ffn, 'w') as fp:
@@ -122,7 +140,8 @@ class FlowIOUtilsFs():
     extension = ext
     
     def _cleaner( fn: str):
-      ffn = f'{fn}.{extension}'
+      ffn = f'{self._config.storage_location}/{fn}'
+      ffn = f'{ffn}.{extension}'
       if os.path.exists (ffn) :
         os.remove (ffn)
       else :
